Remove some of the GetMapInfoAsync methods that are not necessary and merge IRenderInfo into IRenderInfo
