+++ Mapsui.UI.MapView/MapView.cs
@@ -636,11 +636,7 @@ public class MapView : MapControl, INotifyPropertyChanged, IEnumerable<Pin>
 
         if (Map != null)
         {
-            // Check if we hit a drawable/pin/callout etc
-            var getMapInfo = () => GetMapInfo(screenPosition);
-            var getMapInfoAsync = () => GetMapInfoAsync(screenPosition);
-
-            var mapInfoEventArgs = new MapInfoEventArgs(getMapInfo, getMapInfoAsync, e.TapType, handled);
+            var mapInfoEventArgs = new MapInfoEventArgs(() => GetMapInfo(screenPosition), e.TapType, handled);
 
             HandlerInfo(mapInfoEventArgs);
