+++ Mapsui/MapInfoEventArgs.cs
@@ -1,10 +1,12 @@
-﻿using Mapsui.Manipulations;
+﻿using Mapsui.Layers;
+using Mapsui.Manipulations;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Mapsui;
 
-public class MapInfoEventArgs(Func<MapInfo> getMapInfo, Func<Task<MapInfo>> getMapInfoAsync, TapType tapType, bool handled) : EventArgs
+public class MapInfoEventArgs(Func<MapInfo> getMapInfo, TapType tapType, bool handled) : EventArgs
 {
     /// <summary>
     /// Number of times the user tapped the location
@@ -20,5 +22,9 @@ public class MapInfoEventArgs(Func<MapInfo> getMapInfo, Func<Task<MapInfo>> getM
     /// Function to get the map info
     /// </summary>
     public Func<MapInfo> GetMapInfo { get; } = getMapInfo;
-    public Func<Task<MapInfo>> GetMapInfoAsync { get; } = getMapInfoAsync;
+
+    public async Task<MapInfo> GetRemoteMapInfoAsync(ScreenPosition screenPosition, Viewport viewport, IEnumerable<ILayer> layers, int margin)
+    {
+        return await RemoteMapInfoFetcher.GetRemoteMapInfoAsync(screenPosition, viewport, layers, margin);
+    }
 }
