--- Mapsui/Rendering/IRenderInfo.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Mapsui.Layers;
-using Mapsui.Manipulations;
-
-namespace Mapsui.Rendering;
-
-public interface IRenderInfo
-{
-    MapInfo GetMapInfo(ScreenPosition screenPosition, Viewport viewport, IEnumerable<ILayer> layers, int margin = 0);
-    Task<MapInfo> GetRemoteMapInfoAsync(ScreenPosition screenPosition, Viewport viewport, IEnumerable<ILayer> layers, int margin = 0);
-}
